backend/requirements.txt
Flask==3.1.2
flask-cors==6.0.1
opencv-python==4.12.0.88
numpy==2.2.6
scikit-image==0.25.2
scipy==1.15.3
pillow==12.0.0
gunicorn==23.0.0
python-dotenv==1.2.1
imageio==2.37.0
tifffile==2025.5.10
